requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
disnake==2.5.1
idna==3.3
multidict==6.0.2
pip==21.3.1
python-dotenv==0.20.0
setuptools==60.2.0
typing-extensions==4.2.0
wheel==0.37.1
yarl==1.7.2
